feat : Add jwt error handler

+++ controller/LikeController.js
@@ -3,12 +3,16 @@ const { StatusCodes } = require("http-status-codes");
 const { ensureAuthorization } = require("../uitl/auth.js");
 
 const addLike = (req, res) => {
-  const book_id = req.params.id;
+  const bookId = req.params.id;
 
-  const authorization = ensureAuthorization(req, res);
+  const authorization = ensureAuthorization(req);
+
+  if (!authorization.ok) {
+    throw Error(authorization.message);
+  }
 
   let sql = `INSERT INTO likes (user_id, liked_book_id) VALUES (?, ?)`;
-  let values = [authorization.id, book_id];
+  let values = [authorization.id, bookId];
   conn.query(sql, values, (err, results) => {
     if (err) {
       console.log(err);
@@ -20,12 +24,16 @@ const addLike = (req, res) => {
 };
 
 const removeLike = (req, res) => {
-  const book_id = req.params.id;
+  const bookId = req.params.id;
 
-  const authorization = ensureAuthorization(req, res);
+  const authorization = ensureAuthorization(req);
+
+  if (!authorization.ok) {
+    throw Error(authorization.message);
+  }
 
   let sql = `DELETE FROM likes WHERE user_id=? AND liked_book_id=?`;
-  let values = [authorization.id, book_id];
+  let values = [authorization.id, bookId];
 
   conn.query(sql, values, (err, results) => {
     if (err) {
